+++ connecteur/cpp/lib/CPPWrapperExceptionServices.php
@@ -1,0 +1,5 @@
+<?php
+
+class CPPWrapperExceptionServices extends Exception
+{
+}